refactor(dashboard): contain the plugin in BTCPayServer/Plugins/Dashboard
Follow the same systems-plugin layout used by Impersonation,
PointOfSale, etc.: a single directory holds the plugin entry point,
DI registration, models, framework components, widgets, controller
partial, and view. The plugin auto-registers via
BaseBTCPayServerPlugin discovery, so the explicit
services.AddDashboardServices() call in BTCPayServerServices.cs and
the dashboard helpers on BlazorExtensions.cs are no longer needed.

Layout:
  BTCPayServer/Plugins/Dashboard/
  ├── DashboardPlugin.cs              (BaseBTCPayServerPlugin entry)
  ├── _Imports.razor                  (plugin-local Blazor usings)
  ├── BaseWidgetComponent.cs
  ├── DashboardJsInterop.cs
  ├── DashboardPage.razor
  ├── DashboardService.cs
  ├── DashboardWidgetPicker.razor
  ├── DefaultServerDashboardTemplate.cs
  ├── DefaultStoreDashboardTemplate.cs
  ├── IDashboardTemplateProvider.cs
  ├── WidgetConfigPanel.razor
  ├── WidgetContainer.razor
  ├── WidgetRegistry.cs
  ├── WidgetUnavailable.razor
  ├── Controllers/UIStoresController.Dashboard2.cs  (partial; keeps URL)
  ├── Models/{DashboardCollection,DashboardDefinition,
  │           WidgetDescriptor,WidgetPlacement}.cs
  ├── Views/Dashboard2.cshtml
  └── Widgets/NotesWidget.razor

Outside the plugin:
- BTCPayServer.Abstractions/Contracts/IStoreSetupStepProvider.cs
  stays in Abstractions (it's a public extensibility contract).
- BTCPayServer/wwwroot/css/dashboard.css and
  BTCPayServer/wwwroot/js/dashboard-interop.js stay in shared
  wwwroot/ (BTCPay convention — no first-party plugin ships its
  own JS/CSS).
- Gridstack vendor files stay in BTCPayServer/wwwroot/vendor/.

Notes on the controller move:
- The Dashboard2() action is a partial of UIStoresController so the
  URL stays /stores/{storeId}/2. The file lives in the plugin folder
  via the partial-class pattern (same namespace, different folder).
- The view location expander only knows about /Views/UIStores/, so
  the controller returns the absolute view path
  "/Plugins/Dashboard/Views/Dashboard2.cshtml" explicitly.

BlazorExtensions.cs reverts to its master state (just IsPreRendering).
BTCPayServerServices.cs reverts (no AddDashboardServices() call).
Namespace rename: BTCPayServer.Blazor.Dashboard.* -> BTCPayServer.Plugins.Dashboard.*
