+++ BTCPayServer/Plugins/Dashboard/DashboardPage.razor
@@ -1,4 +1,4 @@
-@using BTCPayServer.Blazor.Dashboard.Models
+@using BTCPayServer.Plugins.Dashboard.Models
 @using BTCPayServer.Data
 @using BTCPayServer.Payments
 @using BTCPayServer.Payments.Bitcoin
