+++ BTCPayServer/Plugins/Dashboard/DashboardWidgetPicker.razor
@@ -1,4 +1,4 @@
-@using BTCPayServer.Blazor.Dashboard.Models
+@using BTCPayServer.Plugins.Dashboard.Models
 
 <div class="dropdown">
     <button class="btn btn-outline-primary btn-sm dropdown-toggle" type="button"
