+++ BTCPayServer/Plugins/Dashboard/Models/DashboardDefinition.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace BTCPayServer.Blazor.Dashboard.Models;
+namespace BTCPayServer.Plugins.Dashboard.Models;
 
 public enum DashboardScope
 {
