+++ BTCPayServer/Plugins/Dashboard/Models/WidgetDescriptor.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace BTCPayServer.Blazor.Dashboard.Models;
+namespace BTCPayServer.Plugins.Dashboard.Models;
 
 public enum WidgetScope
 {
