+++ BTCPayServer/Plugins/Dashboard/Views/Dashboard2.cshtml
@@ -1,5 +1,6 @@
-@using BTCPayServer.Blazor.Dashboard
+@using BTCPayServer.Plugins.Dashboard
 @using BTCPayServer.Abstractions.Contracts
+@using BTCPayServer.Views.Stores
 @using Microsoft.AspNetCore.Identity
 @inject IScopeProvider ScopeProvider
 @inject UserManager<BTCPayServer.Data.ApplicationUser> UserManager
