+++ BTCPayServer/Plugins/Dashboard/WidgetConfigPanel.razor
@@ -1,4 +1,4 @@
-@using BTCPayServer.Blazor.Dashboard.Models
+@using BTCPayServer.Plugins.Dashboard.Models
 @using Newtonsoft.Json.Linq
 
 @if (Visible)
