+++ BTCPayServer/Plugins/Dashboard/WidgetContainer.razor
@@ -1,4 +1,4 @@
-@using BTCPayServer.Blazor.Dashboard.Models
+@using BTCPayServer.Plugins.Dashboard.Models
 @using Newtonsoft.Json.Linq
 
 <div class="widget @(Descriptor.CssClass ?? "") @(_featureUnavailable ? "widget-feature-unavailable" : "")">
