+++ BTCPayServer/Plugins/Dashboard/WidgetUnavailable.razor
@@ -1,4 +1,4 @@
-@using BTCPayServer.Blazor.Dashboard.Models
+@using BTCPayServer.Plugins.Dashboard.Models
 
 <div class="widget p-3">
     <h6 class="text-muted">Widget Unavailable</h6>
