+++ BTCPayServer/Plugins/Dashboard/Widgets/NotesWidget.razor
@@ -1,6 +1,6 @@
-@using BTCPayServer.Blazor.Dashboard.Models
+@using BTCPayServer.Plugins.Dashboard.Models
 @using Newtonsoft.Json.Linq
-@inherits BTCPayServer.Blazor.Dashboard.BaseWidgetComponent<NotesWidget.NotesConfig>
+@inherits BTCPayServer.Plugins.Dashboard.BaseWidgetComponent<NotesWidget.NotesConfig>
 
 <header class="mb-2 d-flex justify-content-between align-items-center">
     <h6 class="mb-0">@(TypedConfig?.Title ?? "Notes")</h6>
