+++ BTCPayServer/Plugins/Dashboard/_Imports.razor
@@ -1,0 +1,11 @@
+@using System.Net.Http
+@using Microsoft.AspNetCore.Authorization
+@using Microsoft.AspNetCore.Components.Authorization
+@using Microsoft.AspNetCore.Components.Forms
+@using Microsoft.AspNetCore.Components.Routing
+@using Microsoft.AspNetCore.Components.Web
+@using Microsoft.JSInterop
+@using BTCPayServer.Blazor
+@using BTCPayServer.Abstractions.Extensions
+@using BTCPayServer.Plugins.Dashboard
+@using BTCPayServer.Plugins.Dashboard.Models
